Move ArrayRef, Backtrace, Error, SmallVector, optional to ATen/core; add CoreAPI
