+++ aten/src/ATen/core/ATenCoreTest.h
@@ -1,15 +1,6 @@
 #pragma once
 
-// TODO: Move this to something like ATenCoreGeneral.h
-#ifdef _WIN32
-# if defined(ATen_cpu_EXPORTS) || defined(caffe2_EXPORTS)
-#  define AT_CORE_API __declspec(dllexport)
-# else
-#  define AT_CORE_API __declspec(dllimport)
-# endif
-#else
-# define AT_CORE_API
-#endif
+#include <ATen/core/CoreAPI.h>
 
 namespace at {
